ignore function from coverage
