+++ packages/react-components/lib/tasks/task-panel.tsx
@@ -6,6 +6,7 @@ import * as RmfModels from 'rmf-models';
 import { CreateTaskForm, CreateTaskFormProps } from './create-task';
 import { TaskInfo } from './task-info';
 import { TaskTable } from './task-table';
+import { parseTasksFile } from './utils';
 
 const useStyles = makeStyles((theme) => ({
   detailPanelContainer: {
@@ -34,49 +35,6 @@ function NoSelectedTask() {
 export interface FetchTasksResult {
   tasks: RmfModels.TaskSummary[];
   totalCount: number;
-}
-
-function checkField(obj: Record<string, unknown>, field: string, type: string): void {
-  if (!Object.prototype.hasOwnProperty.call(obj, field) || typeof obj[field] !== type) {
-    throw new TypeError(`expected [${field}] to be [${type}]`);
-  }
-}
-
-function parseTasksFile(contents: string): SubmitTask[] {
-  // TODO: See if we can generate validators from the schema.
-  const tasks = JSON.parse(contents);
-  if (!Array.isArray(tasks)) {
-    throw new TypeError('expected an array');
-  }
-  for (const t of tasks) {
-    if (typeof t !== 'object') {
-      throw new TypeError('expected object');
-    }
-    checkField(t, 'task_type', 'number');
-    checkField(t, 'start_time', 'number');
-    checkField(t, 'priority', 'number');
-    checkField(t, 'description', 'object');
-    const desc = t['description'];
-    switch (t['task_type']) {
-      case RmfModels.TaskType.TYPE_CLEAN:
-        checkField(desc, 'cleaning_zone', 'string');
-        break;
-      case RmfModels.TaskType.TYPE_DELIVERY:
-        checkField(desc, 'pickup_place_name', 'string');
-        checkField(desc, 'pickup_dispenser', 'string');
-        checkField(desc, 'dropoff_ingestor', 'string');
-        checkField(desc, 'dropoff_place_name', 'string');
-        break;
-      case RmfModels.TaskType.TYPE_LOOP:
-        checkField(desc, 'num_loops', 'number');
-        checkField(desc, 'start_name', 'string');
-        checkField(desc, 'finish_name', 'string');
-        break;
-      default:
-        throw new TypeError('unknown task_type');
-    }
-  }
-  return tasks;
 }
 
 export interface TaskPanelProps extends React.HTMLProps<HTMLDivElement> {
